Rename and simplify DataMapper wrappers
- Rename OpenApiEnricherDataMapper → NamedParamsDataMapper (works for any
  service with named typed parameters, not just OpenAPI)
- Delete ParameterDataMapper — inline its trivial logic into parametersListField
- Include complex-typed (non-primitive) params in NamedParamsDataMapper;
  they map to NuType "Any" which skips type-checking but allows mapping
